pyproject.toml
[project]
name = "counter-factual-streetview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3",
    "langgraph>=0.2",
    "openai>=1.43",
    "pillow>=10.4",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "replicate>=0.34",
    "rich>=13.9",
    "typer>=0.12",
]
